+++ packages/element-lib/scripts/couchdb/stop.sh
@@ -1,4 +1,4 @@
-DOCKER_ID=$(docker ps -a -q --filter ancestor=couchdb --format="{{.ID}}")
+DOCKER_ID=$(docker ps -a -q --filter ancestor=couchdb:latest --format="{{.ID}}")
 docker stop $DOCKER_ID > /dev/null
 docker rm $DOCKER_ID > /dev/null
 echo "🏁 COUCHDB Terminated"