+++ packages/element-lib/src/adapters/database/ElementCouchDBAdapter.js
@@ -1,7 +1,7 @@
 const NodeCouchDb = require('node-couchdb');
 
 class ElementCouchDBAdapter {
-  constructor({ name, remote }) {
+  constructor({ name, remote, host }) {
     if (remote) {
       const urlRegex = /https:\/\/(.*):(.*)@(.*)\/(.*)/;
       const parts = urlRegex.exec(remote);
@@ -17,7 +17,15 @@ class ElementCouchDBAdapter {
       });
     } else {
       this.name = name;
-      this.couch = new NodeCouchDb();
+      this.couch = new NodeCouchDb({
+        host,
+        protocol: 'http',
+        port: 5984,
+        auth: {
+          user: 'admin',
+          pass: 'password',
+        },
+      });
     }
   }
 
@@ -74,6 +82,12 @@ class ElementCouchDBAdapter {
     await this.couch.dropDatabase(this.name);
     this.created = false;
     await this.init();
+  }
+
+  async deleteDB() {
+    await this.init();
+    await this.couch.dropDatabase(this.name);
+    this.created = false;
   }
 
   async close() {